Load content information from server in product page part 2

+++ app/src/main/java/com/example/digikala/data/models/product/ProductPageData.kt
@@ -222,7 +222,7 @@ data class Grade(
 )
 
 data class Digiclub(
-    val point: Double,
+    val point: Int,
 )
 
 data class Price2(
